+++ server/main.js
@@ -1,2 +1,2 @@
-import '../imports/api/games.js';
+import '../imports/api/games_api.js';
